Refactors code into organized components and css modules
