+++ .prettierrc
@@ -1,0 +1,14 @@
+{
+    "printWidth": 100,
+    "tabWidth": 2,
+    "useTabs": true,
+    "semi": true,
+    "singleQuote": true,
+    "quoteProps": "as-needed",
+    "jsxSingleQuote": false,
+    "trailingComma": "all",
+    "bracketSpacing": true,
+    "jsxBracketSameLine": false,
+    "endOfLine": "lf",
+    "embeddedLanguageFormatting": "auto"
+}
