+++ src/App.css
@@ -1,0 +1,81 @@
+* {
+  font-family: 'Alata', sans-serif;
+  margin: 0px;
+  padding: 0px;
+  box-sizing: border-box;
+}
+
+body {
+  width: 500px;
+  height: 700px;
+}
+
+
+/* Standard tag styling */
+h1 {
+  font-size: 32px;
+}
+
+pre {
+  resize: vertical;
+  font-family: "Source Code Pro", monospace;
+}
+
+input {
+  width: 100%;
+  height: 24px;
+  padding: 4px;
+  border-radius: 5px;
+  border: none;
+  box-shadow: #E5E5E5 5px 5px;
+  z-index: 2;
+  transition: box-shadow 0.3s ease;
+}
+
+textarea {
+  width: 100%;
+  min-height: 100px;
+  max-height: 300px;
+  height: 100px;
+  border-radius: 5px;
+  box-shadow: #E5E5E5 5px 5px;
+  border: none;
+  padding: 4px;
+  resize: vertical;
+  font-family: "Source Code Pro", monospace;
+  overflow-y: auto;
+  z-index: 2;
+  transition: box-shadow 0.5s ease;
+}
+
+input:focus, textarea:focus {
+  box-shadow: none;
+  outline: none; 
+}
+
+i {
+  cursor: pointer;
+  width: 20px;
+  color: #5941A9;
+}
+
+button {
+  padding: 8px;
+  background-color: #5941A9;
+  color: #f4f4f4;
+  border-radius: 5px;
+  box-shadow: none;
+  border: none;
+  z-index: 2;
+}
+
+button:active {
+  background-color: #9D8AFF;
+}
+
+
+/* Standard variant size of btns, text, etc. [xs, sm, md, lg, xl]  */
+.md-btn {
+  width: 250px;
+  margin-top: 16px;
+}
