+++ src/components/common/bubble_background.jsx
@@ -1,0 +1,12 @@
+import "./bubble_background.css"
+
+export const BubbleBackground = () => {
+    return (
+        <div className="background">
+            <div class="circle circle1"></div>
+            <div class="circle circle2"></div>
+            <div class="circle circle3"></div>
+            <div class="circle circle4"></div>
+        </div>
+    )
+}
