+++ src/components/common/header.jsx
@@ -1,3 +1,5 @@
+import "./header.css"
+
 export const Header = () => {
     return (
         <div className='header'>
