+++ src/components/snippet/snippet.css
@@ -1,0 +1,73 @@
+/* Snippet Column */ 
+.snippet-column {
+    padding-bottom: 8px;
+}
+
+
+/* Snippet Form */
+.form-column {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    row-gap: 24px;
+    z-index: 0;
+    padding: 8px 16px;
+}
+
+.snippet-form {
+    padding: 16px;
+    background-color: black;
+    color: white;
+}
+
+.tm {
+    color: white;
+    display: flex;
+    justify-content: end;
+    font-size: 10px;
+}
+  
+.purple-txt {
+    text-decoration: none;
+    color: #9888D8;
+}
+
+
+/* Snippet List */
+.snippets-container {
+    margin: 16px;
+    padding: 24px;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    row-gap: 16px;
+    border-radius: 8px;
+    border: 3px #5941A9;
+    box-shadow: 2px 2px 10px 1px #BDBDBD;
+}
+
+.snippet {
+    padding: 8px;
+    border: 1px solid #5941A9;
+    border-radius: 5px;
+}
+  
+.snippet h2 {
+    font-size: 16px;
+    max-width: 300px;
+    text-overflow: ellipsis;
+    overflow: hidden;
+    white-space: nowrap;
+}
+  
+.snippet pre {
+    min-height: 36px;
+    max-height: 300px;
+    background: #f4f4f4;
+    margin-top: 4px;
+    padding: 8px;
+    white-space: pre-wrap;
+    overflow-y: auto;
+    overflow-x: auto;
+}
