+++ src/components/snippet/snippet_column.jsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { SnippetForm } from "./snippet_form"
 import { SnippetList } from "./snippet_list"
-import { Header } from "./header"
+import "./snippet.css"
 
 export const SnippetColumn = () => {
     return (
