--- src/styles.css
@@ -1,244 +1,0 @@
-* {
-  font-family: 'Alata', sans-serif;
-  margin: 0px;
-  padding: 0px;
-  box-sizing: border-box;
-}
-
-body {
-  width: 500px;
-  height: 700px;
-}
-
-/* TODO change to class level selectors */
-
-h1 {
-  font-size: 32px;
-}
-
-pre {
-  resize: vertical;
-  font-family: "Source Code Pro", monospace;
-}
-
-form {
-  padding-bottom: 24px;
-}
-
-input {
-  width: 100%;
-  height: 24px;
-  padding: 4px;
-  border-radius: 5px;
-  border: none;
-  box-shadow: #E5E5E5 5px 5px;
-  z-index: 2;
-  transition: box-shadow 0.3s ease;
-}
-
-textarea {
-  width: 100%;
-  min-height: 100px;
-  max-height: 400px;
-  height: 100px;
-  border-radius: 5px;
-  box-shadow: #E5E5E5 5px 5px;
-  border: none;
-  padding: 4px;
-  resize: vertical;
-  font-family: "Source Code Pro", monospace;
-  overflow-y: auto;
-  z-index: 2;
-  transition: box-shadow 0.5s ease;
-}
-
-input:focus, textarea:focus {
-  box-shadow: none;
-  outline: none; 
-}
-
-i {
-  cursor: pointer;
-  width: 20px;
-  color: #5941A9;
-}
-
-button {
-  padding: 8px;
-  background-color: #5941A9;
-  color: #f4f4f4;
-  border-radius: 5px;
-  box-shadow: none;
-  border: none;
-  z-index: 2;
-}
-
-
-
-button:active {
-  background-color: #9D8AFF;
-}
-
-/* Standard size of btns, text, etc. [xs, sm, md, lg, xl]  */
-
-.md-btn {
-  width: 250px;
-  margin-top: 16px;
-}
-
-/* snippet specific */
-
-.snippet {
-  padding: 8px;
-  border: 1px solid #5941A9;
-  border-radius: 5px;
-}
-
-.snippet h2 {
-  font-size: 16px;
-}
-
-.snippet pre {
-  min-height: 24px;
-  max-height: 400px;
-  background: #f4f4f4;
-  padding: 8px;
-  white-space: pre-wrap;
-  overflow-y: auto;
-  overflow-x: auto;
-}
-
-.snippet button {
-  padding-top: 8px;
-  width: 300px;
-}
-
-.form-column {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  row-gap: 24px;
-  z-index: 0;
-  padding: 8px 16px;
-}
-
-.snippets-container {
-  margin: 16px;
-  padding: 24px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  row-gap: 16px;
-  border-radius: 8px;
-  border: 3px #5941A9;
-  box-shadow: 2px 2px 10px 1px #BDBDBD;
-}
-
-
-.snippet-form {
-  padding: 16px;
-  height: 275px;
-  background-color: black;
-  color: white;
-}
-
-/* header specific */
-
-.header {
-  padding: 0px 16px;
-  background-color: black;
-  color: white;
-}
-
-.header-title {
-  position: relative;
-  z-index: 2;
-}
-
-
-/* Circle background */
-
-.circle {
-  position: absolute;
-  right: -100px; /* Start the circles off-screen */
-  width: 50px;
-  height: 50px;
-  border-radius: 50%;
-  z-index: 1;
-  background-color: #9888D8; /* Light blue color */
-  opacity: 0; /* Start invisible */
-  animation: bubble 3s ease-out forwards;
-}
-
-/* Position the circles horizontally at different places */
-.circle1 {
-  left: 60%;
-  top: -5%;
-  width: 225px;
-  height: 225px;
-  animation-delay: 0s;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.circle2 {
-  left: 35%;
-  top: 24%;
-  animation-delay: 0.25s;
-  width: 100px;
-  height: 100px;
-}
-
-.circle3 {
-  left: 15%;
-  top: 10%;
-  animation-delay: 0.5s;
-  width: 125px;
-  height: 125px;
-}
-
-.circle4 {
-  left: -2%;
-  top: 30%;
-  animation-delay: 0.75s;
-  width: 100px;
-  height: 100px;
-}
-/* Animation keyframes */
-@keyframes bubble {
-  0% {
-    transform: translateX(-100vw); /* Start from the bottom */
-    opacity: 0; /* Start invisible */
-  }
-  20% {
-    opacity: 1; /* Fade in */
-  }
-  100% {
-    transform: translateX(0vw); /* Move up 80% of the viewport height */
-    opacity: 1; /* Stay visible */
-  }
-}
-
-.background {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  z-index: 0; 
-  pointer-events: none;
-}
-
-.tm {
-  color: white;
-  display: flex;
-  justify-content: end;
-  font-size: 10px;
-}
-
-.purple-txt {
-  text-decoration: none;
-  color: #9888D8;
-}
